requirements.txt
# ─── Amazon PA API ────────────────────────────────────────
python-amazon-paapi==5.0.0

# ─── OpenAI ───────────────────────────────────────────────
openai==1.30.1

# ─── Agendamento ─────────────────────────────────────────
APScheduler==3.10.4

# ─── HTTP Client ─────────────────────────────────────────
httpx==0.27.0

# ─── Banco de dados ───────────────────────────────────────
SQLAlchemy==2.0.30

# ─── Configuração ─────────────────────────────────────────
python-dotenv==1.0.1

# ─── Logging estruturado ─────────────────────────────────
python-json-logger==2.0.7
